run test DB migration only if yii_test exists

+++ gigareview/up.sh
@@ -55,8 +55,9 @@ docker-compose run --rm console ./database.sh
 
 # Running database migrations
 docker-compose run --rm console ./yii migrate --interactive=0
-docker-compose run --rm console ./yii_test migrate --interactive=0
-
+if [ -f ./yii_test ];then
+  docker-compose run --rm console ./yii_test migrate --interactive=0
+fi
 
 # Launching all the remaining services
 docker-compose up -d public api
